feat(sidebar): 그룹 메뉴에 대한 컴포넌트 분리
